refactor: handler, util을 담는 global 패키지 생성
