+++ src/main/java/webserver/WebServer.java
@@ -5,6 +5,7 @@ import java.net.Socket;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 
+import global.handler.RequestHandler;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
