+++ src/test/java/util/HttpUtilTest.java
@@ -1,5 +1,6 @@
 package util;
 
+import global.util.HttpUtil;
 import org.assertj.core.api.SoftAssertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
